Improve(structure): Abstract atom.subscriptions into a proper file
