+++ lib/index.js
@@ -3,11 +3,10 @@
 
 import React from 'react'
 import { render, unmountComponentAtNode } from 'react-dom'
-import { CompositeDisposable } from 'atom'
 import { store } from './store'
 import { App } from './components/App'
-import { openInputBottomView, closeInputBottomView, setActivePane } from './actions'
 import { log, clear } from './debug'
+import subscriptions from './subscriptions'
 
 const rootDOMId = 'atom-jq-root'
 let rootDOMNode = null
@@ -28,45 +27,7 @@ export default {
       rootDOMNode
     )
 
-    this.subscriptions = new CompositeDisposable()
-
-    const { isPanelVisible } = store.getState()
-
-    this.subscriptions.add(
-      atom.commands.add('atom-workspace', {
-        'atom-jq:open': () => {
-          store.dispatch(openInputBottomView())
-          // store.dispatch(openInputBottomView())
-        },
-        'atom-jq:close': () => {
-          console.log('isPanelVisible', isPanelVisible)
-          if (isPanelVisible) {
-            store.dispatch(closeInputBottomView())
-          }
-        },
-        'core:close': () => {
-          console.log('isPanelVisible', isPanelVisible)
-          if (isPanelVisible) {
-            store.dispatch(closeInputBottomView())
-          }
-        },
-        'core:cancel': () => {
-          console.log('isPanelVisible', isPanelVisible)
-          if (isPanelVisible) {
-            store.dispatch(closeInputBottomView())
-          }
-        }
-      })
-    )
-
-    this.subscriptions.add(
-      atom.workspace.onDidChangeActivePaneItem((paneItem) => {
-        store.dispatch(setActivePane(paneItem))
-      })
-    )
-
-    const activePaneItem = atom.workspace.getActivePaneItem()
-    store.dispatch(setActivePane(activePaneItem))
+    subscriptions.onActivate(store)
   },
 
   deactivate () {
@@ -78,6 +39,6 @@ export default {
       .filter((panel) => panel.className === rootDOMId)
       .map((panel) => panel.destroy())
 
-    this.subscriptions.dispose()
+    subscriptions.onDeactivate(store)
   }
 }
